fix(FIN-033): refactor environment configuration and validation structure
