+++ src/common/config/public-env-config.constant.ts
@@ -1,0 +1,3 @@
+import { validateClientEnv } from '@common/config/client-validator';
+
+export const PublicEnvConfigConstant = validateClientEnv();
