--- src/common/constants/env-config.constant.ts
@@ -1,3 +1,0 @@
-import { validateEnv } from '@common/config/env/validator';
-
-export const EnvConfigConstant = validateEnv();
