+++ src/server/config/env-config.constant.ts
@@ -1,0 +1,3 @@
+import { validateServerEnv } from '@backend/config/server-validator';
+
+export const EnvConfigConstant = validateServerEnv();
