+++ src/server/shared/constant/jwt-secret.constant.ts
@@ -1,3 +1,3 @@
-import { EnvConfigConstant } from '@common/constants/env-config.constant';
+import { EnvConfigConstant } from '@backend/config/env-config.constant';
 
 export const JwtSecretConstant = new TextEncoder().encode(EnvConfigConstant.JWT_SECRET);
